docker_testing/requirements.txt
sentinelsat
pandas
rasterio
matplotlib
scikit-learn
fiona
shapely
ibmcloudant
utm
Flask 
Flask-RESTful 
python-dotenv